Add brightness as setting
