+++ data/settings.csv
@@ -1,5 +1,5 @@
-brightness,156
-bpm,181.3
-mode,3
+brightness,254
+bpm,175.1
+mode,5
